+++ Control_Model_Testing/source/util/memory_management.h
@@ -5,9 +5,9 @@
 #include <stdlib.h>
 
 
-double** memory_allocation_for_storage_arrays(int sim_time, double dt);
+double** memory_allocation_for_storage_arrays(double sim_time, double dt);
 
-double* memory_allocation_for_1_d_arrays(int sim_time, double dt);
+double* memory_allocation_for_1_d_arrays(double sim_time, double dt);
 
 void destroy_arrays(double** array);
